Refactor force-payment and force-report-computed-task: extract common code
