+++ concent_api/api-integration-force-payment.py
@@ -3,30 +3,27 @@
 import os
 import sys
 import time
-from freezegun              import freeze_time
 
-from golem_messages         import message
+from freezegun import freeze_time
 
-from concent_api.settings import CONCENT_PUBLIC_KEY
-from utils.helpers import get_current_utc_timestamp
-from utils.helpers import sign_message
-from utils.testing_helpers import generate_ecc_key_pair
+from golem_messages import message
+
+from utils.helpers import get_current_utc_timestamp, parse_timestamp_to_utc_datetime
 
 from api_testing_common import api_request
 from api_testing_common import count_fails
 from api_testing_common import create_client_auth_message
+from api_testing_common import create_signed_task_to_compute
 from api_testing_common import get_task_id_and_subtask_id
 from api_testing_common import run_tests
+from api_testing_common import PROVIDER_PRIVATE_KEY
+from api_testing_common import REQUESTOR_PRIVATE_KEY
+from api_testing_common import REQUESTOR_PUBLIC_KEY
 from api_testing_common import timestamp_to_isoformat
 
 import requests
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "concent_api.settings")
-
-(PROVIDER_PRIVATE_KEY,  PROVIDER_PUBLIC_KEY)  = generate_ecc_key_pair()
-(REQUESTOR_PRIVATE_KEY, REQUESTOR_PUBLIC_KEY) = generate_ecc_key_pair()
-(DIFFERENT_PROVIDER_PRIVATE_KEY, DIFFERENT_PROVIDER_PUBLIC_KEY)     = generate_ecc_key_pair()
-(DIFFERENT_REQUESTOR_PRIVATE_KEY, DIFFERENT_REQUESTOR_PUBLIC_KEY)   = generate_ecc_key_pair()
 
 
 def force_payment(timestamp = None, subtask_results_accepted_list = None):
@@ -44,39 +41,6 @@ def subtask_results_accepted(timestamp = None, payment_ts = None, task_to_comput
         )
 
 
-def task_to_compute(
-    timestamp                       = None,
-    compute_task_def                = None,
-    provider_public_key             = None,
-    requestor_public_key            = None,
-    requestor_ethereum_public_key   = None,
-    price                           = 1000
-):
-    with freeze_time(timestamp):
-        task_to_compute = message.tasks.TaskToCompute(
-            provider_public_key=provider_public_key if provider_public_key is not None else PROVIDER_PUBLIC_KEY,
-            requestor_public_key=requestor_public_key if requestor_public_key is not None else REQUESTOR_PUBLIC_KEY,
-            compute_task_def = compute_task_def,
-            requestor_ethereum_public_key = requestor_ethereum_public_key,
-            price=price,
-        )
-        sign_message(task_to_compute, REQUESTOR_PRIVATE_KEY)
-        return task_to_compute
-
-
-def compute_task_def(
-    subtask_id  = None,
-    task_id     = None,
-    deadline    = None,
-):
-    compute_task_def = message.tasks.ComputeTaskDef()
-    compute_task_def['subtask_id']  = subtask_id
-    compute_task_def['task_id']     = task_id
-    compute_task_def['deadline']    = deadline
-
-    return compute_task_def
-
-
 @count_fails
 def test_case_2d_send_correct_force_payment(cluster_consts, cluster_url, test_id):
     # Test CASE 2D - Send correct ForcePayment
@@ -86,26 +50,25 @@ def test_case_2d_send_correct_force_payment(cluster_consts, cluster_url, test_id
         subtask_results_accepted_list=[
             subtask_results_accepted(
                 payment_ts=current_time - cluster_consts.payment_due_time - 10,
-                task_to_compute=task_to_compute(
+
+                task_to_compute=create_signed_task_to_compute(
+                    timestamp=parse_timestamp_to_utc_datetime(current_time),
+                    task_id=task_id + 'a',
+                    subtask_id=subtask_id + 'A',
+                    deadline=current_time,
                     requestor_ethereum_public_key="x" * 128,
-                    compute_task_def=compute_task_def(
-                        deadline=current_time,
-                        task_id=task_id + 'a',
-                        subtask_id=task_id + 'A'
-                    )
+                    price=1000,
                 )
             ),
             subtask_results_accepted(
-                timestamp=timestamp_to_isoformat(current_time),
                 payment_ts=current_time - cluster_consts.payment_due_time - 10,
-                task_to_compute=task_to_compute(
-                    timestamp=timestamp_to_isoformat(current_time),
+                task_to_compute=create_signed_task_to_compute(
+                    timestamp=parse_timestamp_to_utc_datetime(current_time),
+                    task_id=task_id + 'b',
+                    subtask_id=subtask_id + 'B',
+                    deadline=current_time,
                     requestor_ethereum_public_key="x" * 128,
-                    compute_task_def=compute_task_def(
-                        deadline=current_time,
-                        task_id=task_id + 'b',
-                        subtask_id=task_id + 'B'
-                    )
+                    price=1000,
                 )
             )
         ]
@@ -157,29 +120,25 @@ def test_case_2c_send_force_payment_with_no_value_to_be_paid(cluster_consts, clu
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'a',
+                        subtask_id=subtask_id + 'A',
+                        deadline=current_time,
                         requestor_ethereum_public_key="x" * 128,
                         price=0,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'a',
-                            subtask_id=task_id + 'A'
-                        )
                     )
                 ),
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'b',
+                        subtask_id=subtask_id + 'B',
+                        deadline=current_time,
                         requestor_ethereum_public_key="x" * 128,
                         price=0,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'b',
-                            subtask_id=task_id + 'B'
-                        )
                     )
                 )
             ]
@@ -210,27 +169,25 @@ def test_case_2b_send_force_payment_beyond_payment_time(cluster_consts, cluster_
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'a',
+                        subtask_id=subtask_id + 'A',
+                        deadline=current_time,
                         requestor_ethereum_public_key="x" * 128,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'a',
-                            subtask_id=task_id + 'A'
-                        )
+                        price=1000,
                     )
                 ),
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'b',
+                        subtask_id=subtask_id + 'B',
+                        deadline=current_time,
                         requestor_ethereum_public_key="x" * 128,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'b',
-                            subtask_id=task_id + 'B'
-                        )
+                        price=1000,
                     )
                 )
             ]
@@ -265,27 +222,25 @@ def test_case_2_a_force_payment_with_subtask_result_accepted_where_ethereum_acco
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time - cluster_consts.payment_due_time - 10,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'a',
+                        subtask_id=subtask_id + 'A',
+                        deadline=current_time,
                         requestor_ethereum_public_key="x" * 128,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'a',
-                            subtask_id=task_id + 'A'
-                        )
+                        price=1000,
                     )
                 ),
                 subtask_results_accepted(
                     timestamp=timestamp_to_isoformat(current_time),
                     payment_ts=current_time - cluster_consts.payment_due_time - 10,
-                    task_to_compute=task_to_compute(
-                        timestamp=timestamp_to_isoformat(current_time),
+                    task_to_compute=create_signed_task_to_compute(
+                        timestamp=parse_timestamp_to_utc_datetime(current_time),
+                        task_id=task_id + 'b',
+                        subtask_id=subtask_id + 'B',
+                        deadline=current_time,
                         requestor_ethereum_public_key="y" * 128,
-                        compute_task_def=compute_task_def(
-                            deadline=current_time,
-                            task_id=task_id + 'b',
-                            subtask_id=task_id + 'B'
-                        )
+                        price=1000,
                     )
                 )
             ]
@@ -302,6 +257,7 @@ def test_case_2_a_force_payment_with_subtask_result_accepted_where_ethereum_acco
 
 if __name__ == '__main__':
     try:
+        from concent_api.settings import CONCENT_PUBLIC_KEY
         run_tests(globals())
     except requests.exceptions.ConnectionError as exception:
         print("\nERROR: Failed connect to the server.\n", file = sys.stderr)
